fix: stabilize io_uring and IOCP backend tests for Linux/Windows CI
- io_uring: user_data test used maxInt(u64) but encoding truncates to 32 bits
- IOCP: fdToHandle() hit unreachable on Windows; fd_t is already HANDLE
- IOCP: NOP submit skips handle association (PostQueuedCompletionStatus only)
- backend: init/deinit test accounts for IOCP returning null fd

+++ src/internal/backend/io_uring.zig
@@ -985,8 +985,9 @@ test "IoUringBackend - user_data preserved" {
     var backend = try IoUringBackend.init(std.testing.allocator, Config{});
     defer backend.deinit();
 
-    // Submit operations with distinct user_data values
-    const user_data_values = [_]u64{ 0xDEADBEEF, 0xCAFEBABE, 0x12345678, std.math.maxInt(u64) };
+    // Submit operations with distinct user_data values (must fit in 32 bits
+    // since the encoding scheme uses upper 32 bits for slab_key)
+    const user_data_values = [_]u64{ 0xDEADBEEF, 0xCAFEBABE, 0x12345678, std.math.maxInt(u32) };
 
     for (user_data_values) |ud| {
         _ = try backend.submit(.{
